WiP (#229)
* wip

* working

* fix: small fixes
