+++ package.json
@@ -39,6 +39,7 @@
     "dotenv-expand": "^10.0.0",
     "fluent-ffmpeg": "^2.1.2",
     "keyv": "^4.5.3",
+    "node-fetch": "^3.3.2",
     "node-schedule": "^2.1.1",
     "openai": "^4.11.1",
     "qrcode": "^1.5.3",
