+++ src/handlers/message/self.ts
@@ -1,7 +1,11 @@
 import { Message } from "whatsapp-web.js";
 import { setStatusFor } from "../../helpers/message";
 import { createContextFromMessage } from "../context";
-import { getCompletionFor, getSources, getSuggestions } from "../completion";
+import {
+  getCompletionWithBing,
+  getSources,
+  getSuggestions,
+} from "../llm-models/completion-bing";
 import { log } from "../../helpers/utils";
 import {
   BOT_PREFIX,
@@ -19,11 +23,15 @@ export async function handleSelfMessage(message: Message) {
   try {
     const context = await createContextFromMessage(message);
 
-    const completion = await getCompletionFor(message, context, streamingReply);
+    const completion = await getCompletionWithBing(
+      message,
+      context,
+      streamingReply
+    );
     let response = completion.response;
 
     if (ENABLE_REMINDERS === "true")
-      response = await handleReminderFor(message, completion);
+      response = await handleReminderFor(message, message.body);
 
     // TODO: must have a way to select them when replying
     // TODO: maybe they can live in a new whatsapp message (sent immediately after the completion)?
